requirements.txt
google-generativeai>=0.7.0
transformers>=4.43.0
torch>=2.0.0
accelerate>=0.27.0
huggingface_hub>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
